// src/model/eventImage.ts
export class EventImage {
  id: number;
  event_id: string;
  name: string;
  description: string;
  image: string = "https://static.wixstatic.com/media/0af972_330aefeed6cb4313b93164437a290661~mv2.jpg/v1/fill/w_800,h_201,al_c,q_80/0af972_330aefeed6cb4313b93164437a290661~mv2.jpg";
  height: number;
  width: number;
  thumbnail: string;
}