// src/model/locations.ts
export class Locations {
  id: number;
  name: string;
  description: string;
  website: string;
  maplink: string;
  email: string;
  mobilenumber: string;
  address1: string;
  address2: string;
  city: string;
  state: string;
  zipcode: string;
  contactmethod: string;
  contactmethod_name: string;
  email_validated: number;
  phone_validated: number;
  address_validated: number;
}